refactor: import model and db modules by their own paths
Delete the `models` and `db` barrels and point every consumer at the file
that defines the symbol. These sit at the bottom of the server chain, so
removing them is a prerequisite for the barrels above.
