--- src/models/index.ts
@@ -1,9 +1,0 @@
-import "server-only";
-export * from "./invitation.model";
-export * from "./feature.model";
-export * from "./guestbook.model";
-export * from "./order.model";
-export * from "./payment.model";
-export * from "./product.model";
-export * from "./review.model";
-export * from "./user.model";
