+++ src/services/premiumFeature.ts
@@ -1,11 +1,11 @@
 import "server-only";
-import type { IFeature } from "@/models";
-import { FeatureModel } from "@/models";
+import type { IFeature } from "@/models/feature.model";
+import { FeatureModel } from "@/models/feature.model";
 import type { PremiumFeatureDto } from "@/core/schemas";
 import type { PremiumFeature } from "@/core/domain";
 import { AppError } from "@/core/domain";
 export type { PremiumFeature } from "@/core/domain";
-import { dbConnect } from "@/db";
+import { dbConnect } from "@/db/connect";
 
 import mongoose from "mongoose";
 import { requireAdmin } from "./auth";
