+++ src/services/user.ts
@@ -3,9 +3,9 @@ import type { Types } from "mongoose";
 import mongoose from "mongoose";
 import type { AdminUserListPage, UserRole } from "@/core/domain";
 import { AppError, DEFAULT_PAGE_SIZE, USER_ROLES } from "@/core/domain";
-import type { BaseUser, IUser } from "@/models";
-import { UserModel } from "@/models";
-import { dbConnect } from "@/db";
+import type { BaseUser, IUser } from "@/models/user.model";
+import { UserModel } from "@/models/user.model";
+import { dbConnect } from "@/db/connect";
 import { hashPassword } from "@/adapters/server/bcrypt";
 import { decrypt, encrypt } from "@/adapters/server/jose";
 import { deleteCookie } from "@/adapters/server/cookies";
