+++ test/integration/app/api/products/route.integration.test.ts
@@ -11,7 +11,7 @@
 import { describe, it, expect, beforeEach, afterAll } from "vitest";
 import { NextRequest } from "next/server";
 import mongoose from "mongoose";
-import { dbConnect } from "@/db";
+import { dbConnect } from "@/db/connect";
 import { buildProductInput, clearCollections } from "@test/support";
 import { createProductService } from "@/services";
 import { GET } from "@/app/api/products/route";
